Simplify (and improve!) sending of fbdata data
This simpler approach avoids problems when memory is reclaimed by the
garbage collector before being sent back.  The `GC.@preserve` used in
the old way preserved the memory while the response object was being
constructed, but not after construction(!), so it could get reclaimed
after construction but before sending, which resulted in garbage being
sent to the client.